api_scanner/functions/vertex_analyze/requirements.txt
google-cloud-pubsub>=2.20.0
vertexai>=1.70.0
google-cloud-aiplatform>=1.66.0

